Move database access functions to asyncy.Database
