+++ tests/unit/Apps.py
@@ -126,7 +126,7 @@ async def test_init_all(patch, magic, async_mock, config, logger, db):
     releases = [
         ['my_app_uuid']
     ]
-    patch.object(Apps, 'get_all_app_uuids_for_deployment',
+    patch.object(Database, 'get_all_app_uuids_for_deployment',
                  return_value=releases)
     patch.object(Apps, 'reload_app', new=async_mock())
 
@@ -193,7 +193,7 @@ async def test_reload_app(patch, config, logger, db, async_mock,
     patch.object(Sentry, 'capture_exc')
     app_logger = magic()
     patch.object(Apps, 'make_logger_for_app', return_value=app_logger)
-    patch.object(Apps, 'update_release_state')
+    patch.object(Database, 'update_release_state')
 
     patch.object(Apps, 'destroy_app', new=async_mock())
     if raise_exc:
@@ -227,22 +227,14 @@ async def test_reload_app(patch, config, logger, db, async_mock,
         logger.error.assert_not_called()
 
     if raise_exc == asyncio_timeout_exc:
-        Apps.update_release_state.assert_called_with(
+        Database.update_release_state.assert_called_with(
             app_logger, config, 'app_id', 'version', ReleaseState.TIMED_OUT)
-
-
-def test_get_all_app_uuids_for_deployment(patch, magic, config):
-    conn = magic()
-    patch.object(Database, 'new_pg_conn', return_value=conn)
-    query = 'select app_uuid from releases group by app_uuid;'
-    ret = Apps.get_all_app_uuids_for_deployment(config)
-    conn.cursor().execute.assert_called_with(query)
-    assert ret == conn.cursor().fetchall()
 
 
 @mark.asyncio
 async def test_deploy_release_many_services(patch):
-    patch.many(Apps, ['make_logger_for_app', 'update_release_state'])
+    patch.object(Apps, 'make_logger_for_app')
+    patch.object(Database, 'update_release_state')
     patch.init(TooManyServices)
 
     stories = {'services': {}}
@@ -254,12 +246,13 @@ async def test_deploy_release_many_services(patch):
                               stories, False, False, 'owner_uuid')
 
     TooManyServices.__init__.assert_called_with(20, 15)
-    Apps.update_release_state.assert_called()
+    Database.update_release_state.assert_called()
 
 
 @mark.asyncio
 async def test_deploy_release_many_apps(patch, magic):
-    patch.many(Apps, ['make_logger_for_app', 'update_release_state'])
+    patch.object(Apps, 'make_logger_for_app')
+    patch.object(Database, 'update_release_state')
     patch.init(TooManyActiveApps)
 
     stories = {'services': {}}
@@ -276,7 +269,7 @@ async def test_deploy_release_many_apps(patch, magic):
                                   stories, False, False, 'owner_uuid')
 
         TooManyActiveApps.__init__.assert_called_with(20, 5)
-        Apps.update_release_state.assert_called()
+        Database.update_release_state.assert_called()
     finally:
         Apps.apps = {}  # Cleanup.
 
@@ -291,7 +284,8 @@ def test_get_app_config(patch):
 
 @mark.asyncio
 async def test_deploy_release_many_volumes(patch, async_mock):
-    patch.many(Apps, ['make_logger_for_app', 'update_release_state'])
+    patch.object(Apps, 'make_logger_for_app')
+    patch.object(Database, 'update_release_state')
     patch.init(TooManyVolumes)
 
     stories = {'services': {}}
@@ -313,7 +307,7 @@ async def test_deploy_release_many_volumes(patch, async_mock):
                               stories, False, False, 'owner_uuid')
 
     TooManyVolumes.__init__.assert_called_with(20, 15)
-    Apps.update_release_state.assert_called()
+    Database.update_release_state.assert_called()
 
 
 @mark.parametrize('raise_exc', [None, exc, asyncy_exc])
@@ -325,7 +319,7 @@ async def test_deploy_release(config, magic, patch, deleted,
     patch.object(Sentry, 'capture_exc')
     patch.object(Kubernetes, 'clean_namespace', new=async_mock())
     patch.object(Containers, 'init', new=async_mock())
-    patch.many(Apps, ['update_release_state'])
+    patch.object(Database, 'update_release_state')
     app_logger = magic()
     patch.object(Apps, 'make_logger_for_app', return_value=app_logger)
     Apps.apps = {}
@@ -346,14 +340,14 @@ async def test_deploy_release(config, magic, patch, deleted,
         {'stories': True}, maintenance, deleted, 'owner_uuid')
 
     if maintenance:
-        assert Apps.update_release_state.call_count == 0
+        assert Database.update_release_state.call_count == 0
         app_logger.warn.assert_called()
     elif deleted:
         app_logger.warn.assert_called()
-        Apps.update_release_state.assert_called_with(
+        Database.update_release_state.assert_called_with(
             app_logger, config, 'app_id', 'version', ReleaseState.NO_DEPLOY)
     else:
-        assert Apps.update_release_state.mock_calls[0] == mock.call(
+        assert Database.update_release_state.mock_calls[0] == mock.call(
             app_logger, config, 'app_id', 'version', ReleaseState.DEPLOYING)
 
         App.__init__.assert_called_with(
@@ -366,10 +360,10 @@ async def test_deploy_release(config, magic, patch, deleted,
             assert Apps.apps.get('app_id') is None
             if raise_exc == exc:
                 Sentry.capture_exc.assert_called()
-            assert Apps.update_release_state.mock_calls[1] == mock.call(
+            assert Database.update_release_state.mock_calls[1] == mock.call(
                 app_logger, config, 'app_id', 'version', ReleaseState.FAILED)
         else:
-            assert Apps.update_release_state.mock_calls[1] == mock.call(
+            assert Database.update_release_state.mock_calls[1] == mock.call(
                 app_logger, config, 'app_id', 'version', ReleaseState.DEPLOYED)
             assert Apps.apps.get('app_id') is not None
 
@@ -379,27 +373,6 @@ def test_make_logger_for_app(patch, config):
     logger = Apps.make_logger_for_app(config, 'my_awesome_app', '17.1')
     logger.start.assert_called()
     logger.adapt.assert_called_with('my_awesome_app', '17.1')
-
-
-@mark.asyncio
-async def test_update_release_state(patch, logger, config):
-    patch.object(Database, 'new_pg_conn')
-    expected_query = 'update releases ' \
-                     'set state = %s ' \
-                     'where app_uuid = %s and id = %s;'
-
-    Apps.update_release_state(logger, config, 'app_id', 'version',
-                              ReleaseState.DEPLOYED)
-
-    Database.new_pg_conn.assert_called_with(config)
-    Database.new_pg_conn.return_value.cursor.assert_called()
-    Database.new_pg_conn.return_value.cursor \
-        .return_value.execute.assert_called_with(
-            expected_query, (ReleaseState.DEPLOYED.value, 'app_id', 'version'))
-
-    Database.new_pg_conn.return_value.commit.assert_called()
-    Database.new_pg_conn.return_value.cursor.return_value.close.assert_called()
-    Database.new_pg_conn.return_value.close.assert_called()
 
 
 @mark.asyncio
@@ -463,7 +436,7 @@ async def test_destroy_app_exc(patch, async_mock, magic,
                                silent, update_db):
     app = magic()
     app.destroy = async_mock(side_effect=exc())
-    patch.object(Apps, 'update_release_state')
+    patch.object(Database, 'update_release_state')
 
     if silent:
         await Apps.destroy_app(app, silent, update_db_state=update_db)
@@ -472,7 +445,7 @@ async def test_destroy_app_exc(patch, async_mock, magic,
             await Apps.destroy_app(app, silent, update_db_state=update_db)
 
     if update_db:
-        assert Apps.update_release_state.mock_calls == [
+        assert Database.update_release_state.mock_calls == [
             mock.call(app.logger, app.config, app.app_id, app.version,
                       ReleaseState.TERMINATING),
             mock.call(app.logger, app.config, app.app_id, app.version,
